file-gestion/backend/requirements.txt
fastapi==0.121.2
uvicorn==0.38.0
pandas==2.3.3
sqlalchemy==2.0.44
python-multipart==0.0.20
asyncpg==0.27.0
python-dotenv==1.0.0
